Add modular inventory components
