--- Assets/Scripts/Inventory/Inventory.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8b7b922c76076f0f31794d0c1e14a791
-timeCreated: 1735689600
