+++ Assets/Scripts/Inventory/InventoryDragHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c19569df5ff77d0dc9ea99c49008c511
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
