+++ Assets/Scripts/Inventory/InventoryModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 95a99f68343ee657e60be014b457f1de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
