+++ Assets/Scripts/Inventory/InventoryTooltip.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10d9ad605cc970af67439b42bc351f5c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
