+++ Assets/Scripts/Inventory/InventoryUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e8db18c792428a8d2c114d1864ed1fb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
